Exported every model to its own file in a models folder
