+++ src/models/book.ts
@@ -1,3 +1,5 @@
+import { LendingRecord } from "./lendingRecord";
+
 export class Book {
   id: string;
   title: string;
@@ -23,36 +25,5 @@ export class Book {
     this.currentLendeeId = currentLendeeId;
     this.dueDate = dueDate;
     this.lendingHistory = lendingHistory;
-  }
-}
-
-export class Customer {
-  id: string;
-  name: string;
-  email: string;
-
-  constructor(id: string, name: string, email: string) {
-    this.id = id;
-    this.name = name;
-    this.email = email;
-  }
-}
-
-export class LendingRecord {
-  customerId: string;
-  lentDate: string;
-  returnDate?: string;
-  dueDate: string;
-
-  constructor(
-    customerId: string,
-    lentDate: string,
-    dueDate: string,
-    returnDate?: string
-  ) {
-    this.customerId = customerId;
-    this.lentDate = lentDate;
-    this.dueDate = dueDate;
-    this.returnDate = returnDate;
   }
 }
