+++ src/models/customer.ts
@@ -1,0 +1,11 @@
+export class Customer {
+  id: string;
+  name: string;
+  email: string;
+
+  constructor(id: string, name: string, email: string) {
+    this.id = id;
+    this.name = name;
+    this.email = email;
+  }
+}
